aggiornamento struttura cartelle per caratteristica e ripristinato App
