+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\index.js":"1","C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\reportWebVitals.js":"2","C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\App.js":"3","C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\componenti.js":"4","C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\home.js":"5","C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\film.js":"6","C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\serietv.js":"7"},{"size":673,"mtime":1616585359622,"results":"8","hashOfConfig":"9"},{"size":362,"mtime":499162500000,"results":"10","hashOfConfig":"9"},{"size":528,"mtime":499162500000,"results":"11","hashOfConfig":"9"},{"size":2012,"mtime":1616585915779,"results":"12","hashOfConfig":"9"},{"size":82,"mtime":1616579615022,"results":"13","hashOfConfig":"9"},{"size":86,"mtime":1616579615055,"results":"14","hashOfConfig":"9"},{"size":93,"mtime":1616579615029,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"m19bmx",{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"27"},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"27"},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"27"},"C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\index.js",[],"C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\reportWebVitals.js",[],["32","33"],"C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\App.js",[],"C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\componenti.js",[],"C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\home.js",[],["34","35"],"C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\film.js",[],"C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\serietv.js",[],{"ruleId":"36","replacedBy":"37"},{"ruleId":"38","replacedBy":"39"},{"ruleId":"36","replacedBy":"40"},{"ruleId":"38","replacedBy":"41"},"no-native-reassign",["42"],"no-negated-in-lhs",["43"],["42"],["43"],"no-global-assign","no-unsafe-negation"]
+[{"C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\index.js":"1","C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\reportWebVitals.js":"2","C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\App.js":"3","C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\genericComponents\\components.js":"4","C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\pageComponents\\tvseries.js":"5","C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\carousel\\carousel.js":"6","C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\pageComponents\\home.js":"7","C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\pageComponents\\film.js":"8"},{"size":581,"mtime":1616595984686,"results":"9","hashOfConfig":"10"},{"size":362,"mtime":499162500000,"results":"11","hashOfConfig":"10"},{"size":599,"mtime":1616595870069,"results":"12","hashOfConfig":"10"},{"size":1913,"mtime":1616596061878,"results":"13","hashOfConfig":"10"},{"size":93,"mtime":1616595796116,"results":"14","hashOfConfig":"10"},{"size":0,"mtime":1616595870051,"results":"15","hashOfConfig":"10"},{"size":167,"mtime":1616595801742,"results":"16","hashOfConfig":"10"},{"size":86,"mtime":1616595809257,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"m19bmx",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\index.js",[],"C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\reportWebVitals.js",[],["35","36"],"C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\App.js",["37"],"C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\genericComponents\\components.js",[],"C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\pageComponents\\tvseries.js",[],"C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\carousel\\carousel.js",[],"C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\pageComponents\\home.js",[],"C:\\Users\\ValentinaD'Orazio\\Desktop\\movie-app\\src\\pageComponents\\film.js",[],{"ruleId":"38","replacedBy":"39"},{"ruleId":"40","replacedBy":"41"},{"ruleId":"42","severity":1,"message":"43","line":1,"column":8,"nodeType":"44","messageId":"45","endLine":1,"endColumn":12},"no-native-reassign",["46"],"no-negated-in-lhs",["47"],"no-unused-vars","'logo' is defined but never used.","Identifier","unusedVar","no-global-assign","no-unsafe-negation"]
