+++ src/App.js
@@ -1,9 +1,11 @@
 import logo from './logo.svg';
 import './App.css';
+import Header from './genericComponents/components'
+
 
 function App() {
   return (
-    <div className="App">
+    /*<div className="App">
       <header className="App-header">
         <img src={logo} className="App-logo" alt="logo" />
         <p>
@@ -18,7 +20,8 @@ function App() {
           Learn React
         </a>
       </header>
-    </div>
+    </div>*/
+    <Header/>
   );
 }
