--- src/home.js
@@ -1,7 +1,0 @@
-
-
-export default function Home(){
-    return(<h1>
-        Home
-    </h1>);
-}
