+++ src/index.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import './css/menucss.css'
-import './css/responsive.css'
-//import App from './App';
-import Header from './componenti'
-//import reportWebVitals from './reportWebVitals';
+import './genericComponents/menucss.css'
+import './responsive/responsive.css'
+import App from './App';
 
-/*ReactDOM.render(
+import reportWebVitals from './reportWebVitals';
+
+ReactDOM.render(
   <React.StrictMode>
     <App />
   </React.StrictMode>,
@@ -18,11 +18,5 @@ import Header from './componenti'
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
-*/
 
-ReactDOM.render(
- 
-    <Header />,
 
-  document.getElementById('root')
-);
