+++ src/pageComponents/home.js
@@ -1,0 +1,12 @@
+
+
+export default function Home() {
+    return (
+        <main>
+            <h1>
+                Home
+            </h1>
+            
+        </main>
+    );
+}
